pyproject.toml
[project]
name = "ipum-lab1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "joblib>=1.5.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pre-commit>=4.2.0",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "pytest-dotenv>=0.5.2",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "uvicorn>=0.34.2",
]
